+++ api/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/python3
+"""Initializer Module"""
