+++ api/v1/views/places.py
